+++ tests/features/test_all_features.py
@@ -2,17 +2,6 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from eyefeatures.features.dist import (
-    DFDist,
-    DTWDist,
-    EucDist,
-    EyeAnalysisDist,
-    HauDist,
-    MannanDist,
-    MultiMatchDist,
-    ScanMatchDist,
-    TDEDist,
-)
 from eyefeatures.features.extractor import Extractor
 from eyefeatures.features.measures import (
     CorrelationDimension,
@@ -26,6 +15,17 @@ from eyefeatures.features.measures import (
     SaccadeUnlikelihood,
     ShannonEntropy,
     SpectralEntropy,
+)
+from eyefeatures.features.scanpath_dist import (
+    DFDist,
+    DTWDist,
+    EucDist,
+    EyeAnalysisDist,
+    HauDist,
+    MannanDist,
+    MultiMatchDist,
+    ScanMatchDist,
+    TDEDist,
 )
 from eyefeatures.features.stats import (
     FixationFeatures,